Adding Players Table
Players table and Stored procedure to insert information about a player
whenever they log on. Will likely copy to userlogoff and / or same gave
as well so we maintain most up to date picture of user

Also added most of the c++ code needed to flag when a user commits
suicide. Will need to be changed to use session id instead of account id
and still need to add call in game code